requirements.txt
torch
gymnasium
gym==0.23.0
gym-minigrid
pufferlib 